修改Long类型基本数据类型的比较

+++ 51fbapi-web/src/main/java/com/ald/fanbei/api/web/api/goods/GetGoodsDetailInfoApi.java
@@ -316,7 +316,7 @@ public class GetGoodsDetailInfoApi implements ApiHandle{
 		final AfResourceDo vipGoodsResourceDo = afResourceService.getConfigByTypesAndSecType(AfResourceType.WEAK_VERIFY_VIP_CONFIG.getCode(), AfResourceSecType.ORDER_WEAK_VERIFY_VIP_CONFIG.getCode());
 		 if (vipGoodsResourceDo != null){
 			 Long vipGoodsId = NumberUtil.objToLongDefault(vipGoodsResourceDo.getValue(), 0L);
-		     if (goods.getRid() == vipGoodsId){
+		     if (goods.getRid().longValue() == vipGoodsId.longValue()){
 		    	 goodsType = AfGoodsSpecType.AUTH.getCode();
 		     }
 		   }
